error boundary component is no longer a modal
