+++ packages/blockchain-wallet-v4-frontend/src/modals/index.js
@@ -7,7 +7,6 @@ import CoinifyExchangeData from './CoinifyExchangeData'
 import CoinifyTradeDetails from './CoinifyTradeDetails'
 import ConfirmDisable2FA from './ConfirmDisable2FA'
 import EditTxDescription from './EditTxDescription'
-import ErrorBoundary from './ErrorBoundary'
 import ExchangeDetails from './ExchangeDetails'
 import ImportBtcAddress from './ImportBtcAddress'
 import MobileNumberChange from './MobileNumberChange'
@@ -49,7 +48,6 @@ const Modals = props => (
     <CoinifyTradeDetails />
     <ConfirmDisable2FA />
     <EditTxDescription />
-    <ErrorBoundary />
     <ExchangeDetails />
     <ImportBtcAddress />
     <MobileNumberChange />
